#16 Added lib directory and moved errors handler and date manipulation into it
