+++ src/lib/dateManipulation.js
